修复read(getint)这种写法引起的bug,头歌测试集open_set全部通过

+++ tests/correct_test/valid_bracket_2.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-double y;
+float y;
 
 int main() {
     y = (((1 + 2) / 3) + 5);
